updoot download script

+++ download.sh
@@ -2,6 +2,14 @@
 
 echo "Downloading $1"
 
-curl -qLf -O "$1"
+FILENAME="$(basename "$1")"
+
+if [ -f "$FILENAME" ]; then
+    echo "File $FILENAME already exists!"
+else
+    echo "Downloading $FILENAME"
+    curl -qLf -O "$1"
+fi
+
 
 echo "Done!"
